Remove redundant metadata editor component
