+++ src/org/labkey/test/pages/query/EditMetadataPage.java
@@ -1,14 +1,13 @@
 package org.labkey.test.pages.query;
 
-import org.labkey.test.BaseWebDriverTest;
 import org.labkey.test.Locator;
 import org.labkey.test.WebDriverWrapper;
 import org.labkey.test.WebTestHelper;
+import org.labkey.test.components.Component;
+import org.labkey.test.components.WebDriverComponent;
 import org.labkey.test.components.bootstrap.ModalDialog;
-import org.labkey.test.components.domain.BaseDomainDesigner;
 import org.labkey.test.components.domain.DomainFormPanel;
-import org.labkey.test.components.react.ReactSelect;
-import org.labkey.test.util.DataRegionTable;
+import org.labkey.test.components.query.AliasFieldDialog;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 
@@ -16,93 +15,110 @@ import java.util.Map;
 
 import static org.labkey.test.WebDriverWrapper.WAIT_FOR_JAVASCRIPT;
 
-public class EditMetadataPage extends BaseDomainDesigner<EditMetadataPage.ElementCache>
+/**
+ * Automates the platform component defined in: query/src/client/QueryMetadataEditor/QueryMetadataEditor.tsx
+ * Generally rendered by "query-metadataQuery.view"
+ */
+public class EditMetadataPage extends WebDriverComponent<EditMetadataPage.ElementCache>
 {
+    private final WebElement el;
+    private final WebDriver driver;
+
     public EditMetadataPage(WebDriver driver)
     {
-        super(driver);
+        this.driver = driver;
+        el = Locator.id("app").findElement(this.driver); // Full page component
     }
 
     public static EditMetadataPage beginAt(WebDriverWrapper webDriverWrapper, String containerPath, String schemaName, String queryName)
     {
         webDriverWrapper.beginAt(WebTestHelper.buildURL("query", containerPath, "metadataQuery",
                 Map.of("schemaName", schemaName, "query.queryName", queryName)));
-        return new EditMetadataPage (webDriverWrapper.getDriver());
-    }
-
-    public DataRegionTable clickViewData(BaseWebDriverTest test, String schemaName, String queryName)
-    {
-        return test.viewQueryData(schemaName, queryName);
-    }
-
-    public SourceQueryPage clickEditSource()
-    {
-        getWrapper().clickAndWait(elementCache().editSourceBtn);
-        return new SourceQueryPage(getDriver());
+        return new EditMetadataPage(webDriverWrapper.getDriver());
     }
 
     @Override
-    public EditMetadataPage clickSave()
+    public WebElement getComponentElement()
     {
-        elementCache().saveBtn.click();
-        return this;
+        return el;
     }
 
-    public String waitForSuccess()
+    @Override
+    public WebDriver getDriver()
     {
-        return Locator.tagWithClass("div", "alert-success").waitForElement(getDriver(), WAIT_FOR_JAVASCRIPT)
-                .getText();
-    }
-
-    public void resetToDefault()
-    {
-        WebDriverWrapper.waitFor(elementCache().resetToDefaultBtn::isEnabled,
-                "The alias button did not become enabled as expected", 2000);
-        elementCache().resetToDefaultBtn.click();
-        new ModalDialog(new ModalDialog.ModalDialogFinder(getDriver()).withTitle("Confirm Reset")).dismiss("Reset");
-    }
-
-    public EditMetadataPage aliasField(String fieldName)
-    {
-        WebDriverWrapper.waitFor(elementCache().aliasFieldBtn::isEnabled,
-                "The alias button did not become enabled as expected", 2000);
-        elementCache().aliasFieldBtn.click();
-        ModalDialog dialog = new ModalDialog(new ModalDialog.ModalDialogFinder(getDriver()).withTitle("Choose a field to wrap"));
-        ReactSelect.finder(getDriver()).findWhenNeeded(dialog).select(fieldName);
-        dialog.dismiss("OK");
-
-        return this;
+        return driver;
     }
 
     public DomainFormPanel fieldsPanel()
     {
-        getWrapper().waitForElementToDisappear(Locator.tagWithText("span", "Loading..."));
-        return elementCache().firstDomainFormPanel;
+        return elementCache().fieldsPanel.expand();
+    }
+
+    public void resetToDefault()
+    {
+        elementCache().resetButton.click();
+        new ModalDialog(new ModalDialog.ModalDialogFinder(getDriver()).withTitle("Confirm Reset")).dismiss("Reset");
+        // TODO: Wait for reset
+    }
+
+    public void viewData()
+    {
+        getWrapper().clickAndWait(elementCache().viewDataButton);
+    }
+
+    public SourceQueryPage clickEditSource()
+    {
+        getWrapper().clickAndWait(elementCache().editSourceButton);
+        return new SourceQueryPage(getDriver());
+    }
+
+    public AliasFieldDialog clickAliasField()
+    {
+        elementCache().aliasFieldButton.click();
+        return new AliasFieldDialog(this);
+    }
+
+    public EditMetadataPage aliasField(String fieldName)
+    {
+        AliasFieldDialog dialog = clickAliasField();
+        dialog.selectAliasField(fieldName);
+        dialog.clickApply();
+
+        return this;
+    }
+
+    public EditMetadataPage clickSave()
+    {
+        elementCache().saveButton.click();
+        WebDriverWrapper.waitFor(()->  Locator.tagWithClass("div", "alert-success")
+                        .containingIgnoreCase("Save Successful").existsIn(this),
+                "Expected success message did not appear in time", WAIT_FOR_JAVASCRIPT);
+        return this;
     }
 
     @Override
-    protected EditMetadataPage.ElementCache newElementCache()
+    protected ElementCache newElementCache()
     {
-        return new EditMetadataPage.ElementCache();
+        return new ElementCache();
     }
 
-    protected class ElementCache extends BaseDomainDesigner<EditMetadataPage.ElementCache>.ElementCache
+    public class ElementCache extends Component<?>.ElementCache
     {
-        protected final WebElement buttonPanel = queryMetadataButtonPanelLocator().findWhenNeeded(this)
-                .withTimeout(WAIT_FOR_JAVASCRIPT);
-        protected final WebElement aliasFieldBtn = Locator.button("Alias Field").findWhenNeeded(buttonPanel).withTimeout(WAIT_FOR_JAVASCRIPT);
-        protected final WebElement editSourceBtn = Locator.button("Edit Source").findWhenNeeded(buttonPanel).withTimeout(WAIT_FOR_JAVASCRIPT);
-        protected final WebElement saveBtn = Locator.button("Save").findWhenNeeded(getDriver()).withTimeout(WAIT_FOR_JAVASCRIPT);
-        protected final WebElement resetToDefaultBtn = Locator.button("Reset To Default").findWhenNeeded(buttonPanel)
-                .withTimeout(WAIT_FOR_JAVASCRIPT);
-        DomainFormPanel firstDomainFormPanel = new DomainFormPanel.DomainFormPanelFinder(getDriver())   // for situations where there's only one on the page
-                .timeout(WAIT_FOR_JAVASCRIPT)
-                .findWhenNeeded(this);
+        protected final DomainFormPanel fieldsPanel = new DomainFormPanel.DomainFormPanelFinder(getDriver())
+                .timeout(1000).findWhenNeeded();
 
-        protected Locator.XPathLocator queryMetadataButtonPanelLocator()
-        {
-            return Locator.byClass("query-metadata-editor-buttons");
-        }
+        protected final WebElement buttonPanel = Locator.byClass("query-metadata-editor-buttons").findWhenNeeded(this);
+
+        private final WebElement resetButton = Locator.button("Reset To Default")
+                .findWhenNeeded(buttonPanel).withTimeout(WAIT_FOR_JAVASCRIPT);
+        private final WebElement aliasFieldButton = Locator.button("Alias Field")
+                .findWhenNeeded(buttonPanel).withTimeout(WAIT_FOR_JAVASCRIPT);
+        private final WebElement viewDataButton = Locator.button("View Data")
+                .findWhenNeeded(buttonPanel).withTimeout(WAIT_FOR_JAVASCRIPT);
+        private final WebElement editSourceButton = Locator.button("Edit Source")
+                .findWhenNeeded(buttonPanel).withTimeout(WAIT_FOR_JAVASCRIPT);
+        public final WebElement saveButton = Locator.byClass("save-button")
+                .findWhenNeeded(buttonPanel).withTimeout(WAIT_FOR_JAVASCRIPT);
 
     }
 }
